Deprecated functions of quantumPPMS

get.vsm.sweepData,ppms.removeSubstrateMagnetization and ppms.getSusceptibility are depreacted, so avoid using this or downgrade to an older function; these functions may disappear from this package

get.vsm.sweepData(obj)

ppms.getSusceptibility(h)

ppms.removeSubstrateMagnetization(hyst)

ppms.seqType(Temp, Happ)

ppms.vsm.hystLoops(H, M)

vsm.get.HystLoops(data)

vsm.stats(hyst)

Arguments

obj

PPMSdata object

Temp

temperature vector (K)

Happ

applied field vector (Oe)

H

applied magnetic field (Oe)

M

magnetization (emu)

data

data frame with time, T, H, M, Merr

Details

'get.vsm.sweepData' Return only data of magnetic sweeps (hyst loop)

'ppms.seqType': Returns characteristics of a PPMS data file.

'ppms.vsm.hystLoops': Returns a vector that separates different hysteresis loops 'vsm.get.HystLoops' extract hysteresis loops (OBSOLETE, use ppms.vsm.hystLoops)

use summary(PPMSdata) instead

The category number can be used to check whether the same sequence was used to generated the data.