Statistics for one Hysteresis Loop

Retrieve statistics for one loop, if you need to get statistics for all loops in the VSMdata object, then use vsm.hystStat() instead

Retrieve statistics for one loop, if you need to get statistics for all loops in the VSMdata object, then use vsm.hystStat() instead

vsm.hystStatsLoop(obj, loop = 1, direction = 1)

vsm.hystStatsLoop(obj, loop = 1, direction = 1)

Arguments

obj

VSMdata object

loop

number of the loop

direction

+1 or -1 for direction of applied field

Value

list or NULL if not a "MvsH" type

list or NULL if not a "MvsH" type

Examples

filename = vsm.getSampleFiles()[1]
d = vsm.import(filename)
vsm.hystStatsLoop(d)
#> $H.first
#> [1] -19852.77
#> 
#> $H.last
#> [1] 19999.11
#> 
#> $H.max
#> [1] 19999.11
#> 
#> $H.min
#> [1] -19852.77
#> 
#> $Ms1
#> [1] 8.630711e-05
#> 
#> $Ms1.sd
#> [1] 3.463391e-07
#> 
#> $Ms2
#> [1] -8.633927e-05
#> 
#> $Ms2.sd
#> [1] 5.919962e-07
#> 
#> $Hc
#> [1] 413.6
#> 
#> $Hc.err
#> [1] 76.4
#> 
#> $Hc.nFit
#> [1] 7
#> 
#> $Mrem
#> [1] -3.901093e-05
#> 
#> $Mrem.sd
#> [1] 1.778054e-06
#> 
#> $Susceptibility
#> [1] -4.879e-09
#> 
#> $Susceptibility.sd
#> [1] 3.171e-11
#> 
#> $T
#> [1] 2.999695
#> 
#> $T.sd
#> [1] 0.001849705
#> 
#> $speed
#> [1] 104.4547
#> 
#> $speed.sd
#> [1] 0.093174
#> 
#> $time.delta
#> [1] 397.88
#> 
#> $data.points
#> [1] 263
#> 
#> $dir
#> [1] 1
#> 
#> $type
#> [1] "MvsH"
#> 
#> $loop
#> [1] 1
#> 

filename = vsm.getSampleFiles()[1]
d = vsm.import(filename)
vsm.hystStatsLoop(d)
#> $H.first
#> [1] -19852.77
#> 
#> $H.last
#> [1] 19999.11
#> 
#> $H.max
#> [1] 19999.11
#> 
#> $H.min
#> [1] -19852.77
#> 
#> $Ms1
#> [1] 8.630711e-05
#> 
#> $Ms1.sd
#> [1] 3.463391e-07
#> 
#> $Ms2
#> [1] -8.633927e-05
#> 
#> $Ms2.sd
#> [1] 5.919962e-07
#> 
#> $Hc
#> [1] 413.6
#> 
#> $Hc.err
#> [1] 76.4
#> 
#> $Hc.nFit
#> [1] 7
#> 
#> $Mrem
#> [1] -3.901093e-05
#> 
#> $Mrem.sd
#> [1] 1.778054e-06
#> 
#> $Susceptibility
#> [1] -4.879e-09
#> 
#> $Susceptibility.sd
#> [1] 3.171e-11
#> 
#> $T
#> [1] 2.999695
#> 
#> $T.sd
#> [1] 0.001849705
#> 
#> $speed
#> [1] 104.4547
#> 
#> $speed.sd
#> [1] 0.093174
#> 
#> $time.delta
#> [1] 397.88
#> 
#> $data.points
#> [1] 263
#> 
#> $dir
#> [1] 1
#> 
#> $type
#> [1] "MvsH"
#> 
#> $loop
#> [1] 1
#>